Define the functionalities of EmployeeService
I added a dependency injection of the class HttpClient and use that to request and http request on the backend.
Created a variable string which is the resource link of the backend. Created a public method getAllEmployees and use the object
http to perform a HTTP GET request. As of now it was set into <Any> rerturn to any type but later will change that in specific way.